Add use of static_cast to get past Tizen and ASR build issues.

+++ src/app/clusters/camera-av-settings-user-level-management-server/camera-av-settings-user-level-management-server.cpp
@@ -218,7 +218,7 @@ void CameraAvSettingsUserLevelMgmtServer::UpdatePresetID()
         });
         if (it == mMptzPresetHelper.end())
         {
-            currentPresetID = currentPresetID + increment;
+            currentPresetID = static_cast<uint8_t>(currentPresetID + increment);
             break;
         }
         increment++;
@@ -578,7 +578,7 @@ void CameraAvSettingsUserLevelMgmtServer::HandleMPTZRelativeMove(HandlerContext 
             return;
         }
         int16_t panDeltaValue = panDelta.Value();
-        newPan                = newPan + panDeltaValue;
+        newPan                = static_cast<int16_t>(newPan + panDeltaValue);
         if (newPan > mPanMax)
         {
             newPan = mPanMax;
@@ -600,7 +600,7 @@ void CameraAvSettingsUserLevelMgmtServer::HandleMPTZRelativeMove(HandlerContext 
             return;
         }
         int16_t tiltDeltaValue = tiltDelta.Value();
-        newTilt                = newTilt + tiltDeltaValue;
+        newTilt                = static_cast<int16_t>(newTilt + tiltDeltaValue);
         if (newTilt > mTiltMax)
         {
             newTilt = mTiltMax;
@@ -622,7 +622,7 @@ void CameraAvSettingsUserLevelMgmtServer::HandleMPTZRelativeMove(HandlerContext 
             return;
         }
         int8_t zoomDeltaValue = zoomDelta.Value();
-        newZoom               = newZoom + zoomDeltaValue;
+        newZoom               = static_cast<int8_t>(newZoom + zoomDeltaValue);
         if (newZoom > mZoomMax)
         {
             newZoom = mZoomMax;
